add_user_id_to_tasks_table.phpを編集

+++ database/migrations/2024_07_19_090834_add_user_id_to_tasks_table.php
@@ -9,19 +9,22 @@ class AddUserIdToTasksTable extends Migration
     public function up()
     {
         Schema::table('tasks', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->after('id');
+            if (!Schema::hasColumn('tasks', 'user_id')) {
+                $table->unsignedBigInteger('user_id')->after('id');
 
-            // 外部キー制約を追加
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                // 外部キー制約を追加
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            }
         });
     }
 
     public function down()
     {
         Schema::table('tasks', function (Blueprint $table) {
-            $table->dropForeign(['user_id']);
-            $table->dropColumn('user_id');
+            if (Schema::hasColumn('tasks', 'user_id')) {
+                $table->dropForeign(['user_id']);
+                $table->dropColumn('user_id');
+            }
         });
     }
 }
-
